backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
httpx==0.27.0
anthropic==0.25.1
transformers
torch
scipy
scikit-learn
python-dotenv
